requirements.txt
# Core dependencies for LSTM Sentiment Classifier

# Deep Learning
torch>=1.9.0

# Scientific Computing
numpy>=1.21.0

# Machine Learning
scikit-learn>=1.0.0

# Data Processing
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration
PyYAML>=5.4.0

# Progress Bars
tqdm>=4.62.0

# Utilities
requests>=2.25.0

# API (if using run_api.py)
fastapi
uvicorn

# Optional NLP Libraries for Advanced Preprocessing
# Install with: pip install nltk spacy
# For spaCy: python -m spacy download en_core_web_sm
# nltk>=3.6.0
# spacy>=3.4.0